refactor:근로관리 중복 코드 삭제
