+++ src/components/WorkManagement/Commute/CommuteTable.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { useMemo } from 'react';
 
 import { GenderIcon } from '@/components/WorkManagement/Work/WorkTableRows';
-import { commuteMockData } from '@/constants/workManagementTable/workTable.mock';
+import { commuteMockData } from '@/constants/workManagement/workTable.mock';
 import { ICommuteData } from '@/models/work.model';
 
 export default function CommuteTable({
