+++ src/components/WorkManagement/Work/WorkTable.tsx
@@ -1,8 +1,8 @@
 import { Table } from '@radix-ui/themes';
 
 import WorkTableRows from '@/components/WorkManagement/Work/WorkTableRows';
-import { WORKTABLE } from '@/constants/workManagementTable/workTable';
-import { workMockData } from '@/constants/workManagementTable/workTable.mock';
+import { WORKTABLE } from '@/constants/workManagement/workTable';
+import { workMockData } from '@/constants/workManagement/workTable.mock';
 
 export default function WorkTable() {
   return (
