+++ src/components/WorkManagement/WorkHeader.tsx
@@ -1,44 +1,19 @@
 import { Button } from '@radix-ui/themes';
 import dayjs, { Dayjs } from 'dayjs';
-import React from 'react';
 import { useMemo } from 'react';
 import { PiGear } from 'react-icons/pi';
-import { useNavigate } from 'react-router-dom';
 
 import { ChangeMonth } from '@/components/Common/ChangeMonth';
-import { ItabItem } from '@/pages/workManagement/WorkManagementPage';
+import { TAB_ITEMS, TabItem } from '@/constants/workManagement/workManagement';
 
 export interface IProps {
-  selectTab: ItabItem;
-  setSelectTab: React.Dispatch<React.SetStateAction<ItabItem>>;
+  selectTab: TabItem;
+  onTabChange: (item: TabItem) => void;
   currentDate: Dayjs;
-  setCurrentDate: React.Dispatch<React.SetStateAction<Dayjs>>;
+  onChangeMonth: (newDate: Dayjs) => void;
 }
 
-export default function WorkHeader({
-  selectTab,
-  setSelectTab,
-  currentDate,
-  setCurrentDate,
-}: IProps) {
-  const navigate = useNavigate();
-
-  const tabItems: ItabItem[] = [
-    { id: 0, title: '근로 관리', path: '/work-management/working' },
-    { id: 1, title: '파트타이머 관리', path: '/work-management/partTime' },
-    { id: 2, title: '출퇴근 관리', path: '/work-management/commute' },
-  ];
-
-  const handleChangeMonth = (newDate: dayjs.Dayjs) => {
-    setCurrentDate(newDate);
-    // backend API 연결
-  };
-
-  const handleTabChange = (item: ItabItem) => {
-    setSelectTab(item);
-    navigate(item.path);
-  };
-
+export default function WorkHeader({ selectTab, currentDate, onTabChange, onChangeMonth }: IProps) {
   const date = useMemo(() => {
     const isCurrentMonth =
       currentDate.month() === dayjs().month() && currentDate.year() === dayjs().year();
@@ -70,20 +45,18 @@ export default function WorkHeader({
       <div className="text-2xl font-medium flex gap-3">
         {selectTab.title}
         <div className="flex gap-3">
-          {tabItems
-            .filter((item) => item.id !== selectTab.id)
-            .map((item) => (
-              <Button
-                key={item.id}
-                color="gray"
-                variant="soft"
-                radius="full"
-                onClick={() => handleTabChange(item)}
-              >
-                <PiGear />
-                <span className="-mb-px mr-1">{item.title}</span>
-              </Button>
-            ))}
+          {TAB_ITEMS.filter((item) => item.id !== selectTab.id).map((item) => (
+            <Button
+              key={item.id}
+              color="gray"
+              variant="soft"
+              radius="full"
+              onClick={() => onTabChange(item)}
+            >
+              <PiGear />
+              <span className="-mb-px mr-1">{item.title}</span>
+            </Button>
+          ))}
         </div>
       </div>
       <div className="flex items-center	gap-2">
@@ -92,7 +65,7 @@ export default function WorkHeader({
           <span className="font-bold text-black">{date.workingDays}</span>일 (데이터기준
           {date.period})
         </p>
-        <ChangeMonth currMonth={currentDate} onChangeMonth={handleChangeMonth} />
+        <ChangeMonth currMonth={currentDate} onChangeMonth={onChangeMonth} />
       </div>
     </div>
   );
