+++ src/components/WorkManagement/WorkSelect.tsx
@@ -1,10 +1,7 @@
 import { useForm } from 'react-hook-form';
 
 import SelectBox from '@/components/Common/Select';
-import {
-  branchMockData,
-  departmentMockData,
-} from '@/constants/workManagementTable/workSelect.mock';
+import { branchMockData, departmentMockData } from '@/constants/workManagement/workSelect.mock';
 
 export default function WorkSelect() {
   const { register } = useForm();
